reftests: Allow minor differences to be tolerated
Based on an earlier patch by Michael Biebl, as used in Debian's GTK 3
packaging, with additional inspiration from librsvg's reftests.

Each .ui or .node reftest can have an accompanying .keyfile file
like this:

    [reftest]
    accepted-diff-level=2
    accepted-diff-pixels=100
    tolerated-diff-level=20
    tolerated-diff-pixels=1000

If the number of pixels that differ from the reference is no more than
accepted-diff-pixels, and each channel in each of those pixels differs
from the reference by no more than accepted-diff-level, then we consider
that to be a full success, and don't even save the .diff.png for
analysis.

If that check fails, but the number of pixels that differ is no more
than tolerated-diff-pixels and the differences are no more than
tolerated-diff-level, then we treat it as a success with warnings, save
the .diff.png for analysis, and use g_test_incomplete() to record the
test-case as "TODO".

Signed-off-by: Simon McVittie <smcv@debian.org>