gskngltexturelibraryprivate.h: Fix build on Visual Studio
The 'GskNglTextureAtlasEntry' structure is used as a member in two other
structures in two Gsk private headers, namely 'GskNglGlyphValue' and
'GskNglIconData', but the build breaks on Visual Studio since
'GskNglTextureAtlasEntry' contains a last member that is a zero-sized array
that is normally OK if it is the last member of structure -or- if the structure
that contains a 'GskNglTextureAtlasEntry' structure as its last member.
Otherwise, a GCCism is involved here.

Fix the build on Visual Studio compilers by not using a zero-sized array as
a last member of 'GskNglTextureAtlasEntry', on Visual Studio compilers.