GDK-Win32: Remove extraneous call to _gdk_win32_append_event()
Apparently, by comparing with the other backends, we should not call
_gdk_win32_append_event() after calling gdk_scroll_event_new() but we should
call it after calling gdk_scroll_event_new_discrete(), which was why we didn't
restore the cursor after we scroll using the mouse wheel and didn't manage to
remove the shade that appears after we scrolled to the very top or very bottom.

Also use IDC_SIZEALL for the system cursor that we fall back to if no cursor
theme is installed, as with other Windows programs, suggested by the reporter
of issue #3581.

Fixes issue #3581.