build: Define if options should be enabled or disbaled
"auto" options are terrible because nobody knows if they should enable
them or not. So enable the ones that every release should use and
disable the ones that releases aren't expected to use.

Everybody is still free to override this in their config - either to
turn off stuff that's unsupported or not wanted for them or to enable
new and exciting experimental features - but we should not leave this to
the randomly existing environment that somebpody happens to build in.

Options should be a deliberate choice every time.