gtk: add GtkTextRegion
GtkTextRegion is a helper for tracking regions of text when you might not
be able to use GtkTextBuffer's B-Tree. This can be useful when writing
code that needs to work with GtkEditable and GtkTextView both.

This is implemented in a fashion that resembles the combination of a
B+Tree (doubly-linked internal and leaf nodes) and a piecetable.

The goal for this is to be able to track regions of a buffer that need
updating as we process GtkTextView drawing.

A number of tests are provided to ensure correctness of the data structure.