textbuffer: add rudimentary spell checking
This only checks spelling and does not yet hook up into anything such as
suggestions. It must be driven by the display to force checking a visible
region. To enable, set the :spell-checker property.

Longer term we'll want something more advanced than the simple iteration
through words that is done here. We need to have regions that are not to
be spell checked so that GtkSourceView can continue to work as expected.

We also need to apply the correct language from runs of text which would
benefit from being connected to a language (and PangoLanguage may not be
suitable there).